+++ src/main/java/gyro/google/compute/BackendServiceResource.java
@@ -25,12 +25,14 @@ import java.util.stream.Collectors;
 import com.google.api.services.compute.Compute;
 import com.google.api.services.compute.model.BackendService;
 import com.google.api.services.compute.model.Operation;
+import com.google.api.services.compute.model.SecurityPolicyReference;
 import com.google.cloud.compute.v1.ProjectGlobalBackendServiceName;
 import gyro.core.GyroUI;
 import gyro.core.Type;
 import gyro.core.resource.Resource;
 import gyro.core.resource.Updatable;
 import gyro.core.scope.State;
+import gyro.core.validation.ValidationError;
 
 /**
  * Creates a backend service.
@@ -50,13 +52,15 @@ import gyro.core.scope.State;
  *
  *         health-check: [ $(google::compute-health-check health-check-example-backend-service) ]
  *
+ *         security-policy: $(google::compute-security-policy security-policy-example-backend-service)
+ *
  *         connection-draining
  *             draining-timeout-sec: 30
  *         end
  *
  *         load-balancing-scheme: "EXTERNAL"
  *
- *         enable-cdn: true
+ *         enable-cdn: false
  *         protocol: "HTTPS"
  *         session-affinity: "NONE"
  *         port-name: "http"
@@ -81,6 +85,7 @@ public class BackendServiceResource extends AbstractBackendServiceResource {
 
     private List<BackendSignedUrlKey> signedUrlKey;
     private String portName;
+    private SecurityPolicyResource securityPolicy;
 
     /**
      * Signed Url key configuration for the backend service.
@@ -110,10 +115,26 @@ public class BackendServiceResource extends AbstractBackendServiceResource {
         this.portName = portName;
     }
 
+    /**
+     * The security policy associated with this backend service. This can only be added when ``enableCdn`` is ``false``.
+     */
+    @Updatable
+    public SecurityPolicyResource getSecurityPolicy() {
+        return securityPolicy;
+    }
+
+    public void setSecurityPolicy(SecurityPolicyResource securityPolicy) {
+        this.securityPolicy = securityPolicy;
+    }
+
     @Override
     public void copyFrom(BackendService model) {
         super.copyFrom(model);
+
         setPortName(model.getPortName());
+        setSecurityPolicy(model.getSecurityPolicy() != null
+            ? findById(SecurityPolicyResource.class, model.getSecurityPolicy())
+            : null);
 
         if (getCdnPolicy() != null) {
             // add any new keys not configured through gyro
@@ -153,6 +174,12 @@ public class BackendServiceResource extends AbstractBackendServiceResource {
         Operation operation = client.backendServices().insert(getProjectId(), backendService).execute();
         waitForCompletion(client, operation);
 
+        if (getSecurityPolicy() != null) {
+            state.save();
+
+            saveSecurityPolicy(client);
+        }
+
         if (!getSignedUrlKey().isEmpty()) {
             state.save();
 
@@ -174,8 +201,16 @@ public class BackendServiceResource extends AbstractBackendServiceResource {
 
         Compute client = createComputeClient();
 
-        BackendService backendService = getBackendService(changedFieldNames);
+        BackendServiceResource currentBackendResource = (BackendServiceResource) current;
+        boolean securityPolicyUpdated = false;
 
+        if (changedFieldNames.contains("enable-cdn") && getEnableCdn()
+            && currentBackendResource.getSecurityPolicy() != null) {
+            saveSecurityPolicy(client);
+            securityPolicyUpdated = true;
+        }
+
+        BackendService backendService = getBackendService(changedFieldNames);
         Operation operation = client.backendServices().patch(getProjectId(), getName(), backendService).execute();
         waitForCompletion(client, operation);
 
@@ -185,7 +220,7 @@ public class BackendServiceResource extends AbstractBackendServiceResource {
 
         if (changedFieldNames.contains("signed-url-key")) {
             // delete old keys
-            List<String> deleteSignedUrlKeys = ((BackendServiceResource) current).getSignedUrlKey().stream().map(
+            List<String> deleteSignedUrlKeys = currentBackendResource.getSignedUrlKey().stream().map(
                 BackendSignedUrlKey::getKey).collect(
                 Collectors.toList());
 
@@ -204,6 +239,10 @@ public class BackendServiceResource extends AbstractBackendServiceResource {
                         .execute());
             }
         }
+
+        if (changedFieldNames.contains("security-policy") && !securityPolicyUpdated) {
+            saveSecurityPolicy(client);
+        }
     }
 
     @Override
@@ -211,6 +250,34 @@ public class BackendServiceResource extends AbstractBackendServiceResource {
         Compute client = createComputeClient();
         Operation response = client.backendServices().delete(getProjectId(), getName()).execute();
         waitForCompletion(client, response);
+    }
+
+    private void saveSecurityPolicy(Compute client) throws Exception {
+        SecurityPolicyResource securityPolicyResource = getSecurityPolicy();
+        SecurityPolicyReference securityPolicyReference = null;
+
+        if (securityPolicyResource != null) {
+            securityPolicyReference = new SecurityPolicyReference();
+            securityPolicyReference.set("securityPolicy", securityPolicyResource.getSelfLink());
+        }
+
+        Operation securityPolicyOperation = client.backendServices()
+            .setSecurityPolicy(getProjectId(), getName(), securityPolicyReference)
+            .execute();
+        waitForCompletion(client, securityPolicyOperation);
+    }
+
+    @Override
+    public List<ValidationError> validate(Set<String> configuredFields) {
+        List<ValidationError> errors = new ArrayList<>();
+        if (getEnableCdn() && getSecurityPolicy() != null) {
+            errors.add(new ValidationError(
+                this,
+                "enable-cdn",
+                "'enable-cdn' can't be true when a security policy is provided."));
+        }
+
+        return errors;
     }
 
     static boolean isBackendService(String selfLink) {